check_for_xs: omit cheching in 'win32' directories
Rationale: Windows binary laucners are of no interest to us as we don't build
them.

In theory, the C files there could be part of the module even on non-Windows
(for example by implementing communication with remote windows processes), but
it is unlikely that these would be the only C files so arch detecthin should
still work.