TODO -= possible use of system-wide DhMakePerl/dh-make-perl in tests
dh-make-perl is run from a location relative to the test itself, so it is not
the system-wide.
as for the modules, the tests are run with PERL5LIB set to the working copy (or
some staging directory Module::Build have set up) and PERL5LIB variable is
exported via the system() call.
as a confirmation, running the tests locally (not in a chroot) after changing
the default debhelper level to 8 in DhMakePerl::Config (and in
t/dists/*/wanted-debian*/{control,compat}) shows no problems even if the
system-wide dh-make-perl still defaults to dh7.