Code-deduplication by splitting up guess_tarball()
The same guessing of tar ball names was done in guess_tarball() as
well as in setup_dir(). The common parts are now in guess_tarball()
which needs a prefix (i.e. everything without .tar.gz) as parameter.

The original use of guess_tarball() in execute() has been replaced by
guess_debian_tarball() which just computes the proper prefix and then
calls guess_tarball() on it.

This facilitates later changes on the algorithm to guess tarballs.